Add CardStats, CardType, and CardVariant classes for card management - von julian
