+++ src/WaterWizard.Shared/CardType.cs
@@ -1,0 +1,9 @@
+namespace WaterWizard.Shared;
+
+public enum CardType
+{
+    Damage,
+    Utility,
+    Environment,
+    Healing
+}
