+++ src/WaterWizard.Shared/CardVariant.cs
@@ -1,0 +1,34 @@
+namespace WaterWizard.Shared;
+
+public enum CardVariant
+{
+    // Damage Variants
+    ArcaneMissile,
+    Firebolt,
+    Fireball,
+    GreedHit,
+    FrostBolt,
+
+    // Utility Variants
+    HoveringEye,
+    SummonShip,
+    Teleport,
+    Paralize,
+    ConeOfCold,
+    MinorIllusion,
+    Polymorph,
+
+    // Environment Variants
+    Thunder,
+    Storm,
+    SpawnRocks,
+    RiseSun,
+    CallWind,
+
+    // Healing Variants
+    Heal,
+    Mending,
+    MassMending,
+    PerfectMending,
+    Lifesteal
+}
